Extract shared push token format module for frontend/backend parity
