+++ _api/push/_shared.ts
@@ -1,4 +1,9 @@
 import type { IncomingHttpHeaders } from "node:http";
+import {
+  isPushTokenFormat,
+  PUSH_TOKEN_MAX_LENGTH,
+  PUSH_TOKEN_MIN_LENGTH,
+} from "../../shared/pushToken.js";
 
 export type PushPlatform = "ios" | "android";
 
@@ -9,11 +14,7 @@ export interface PushTokenMetadata {
 }
 
 export const PUSH_TOKEN_TTL_SECONDS = 60 * 60 * 24 * 365;
-export const PUSH_TOKEN_MIN_LENGTH = 20;
-export const PUSH_TOKEN_MAX_LENGTH = 512;
-const PUSH_TOKEN_REGEX = new RegExp(
-  `^[A-Za-z0-9:_\\-.]{${PUSH_TOKEN_MIN_LENGTH},${PUSH_TOKEN_MAX_LENGTH}}$`
-);
+export { PUSH_TOKEN_MIN_LENGTH, PUSH_TOKEN_MAX_LENGTH };
 
 export function isPlainObject(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
@@ -90,7 +91,7 @@ export function getPushTokenSuffix(token: string, suffixLength: number = 8): str
 }
 
 export function isValidPushToken(token: string): boolean {
-  return PUSH_TOKEN_REGEX.test(token);
+  return isPushTokenFormat(token);
 }
 
 export function isPushPlatform(platform: string): platform is PushPlatform {
