+++ src/utils/pushToken.ts
@@ -1,19 +1,11 @@
-export const PUSH_TOKEN_MIN_LENGTH = 20;
-export const PUSH_TOKEN_MAX_LENGTH = 512;
+import {
+  normalizePushTokenValue,
+  PUSH_TOKEN_MAX_LENGTH,
+  PUSH_TOKEN_MIN_LENGTH,
+} from "../../shared/pushToken";
 
-const PUSH_TOKEN_FORMAT_REGEX = new RegExp(
-  `^[A-Za-z0-9:_\\-.]{${PUSH_TOKEN_MIN_LENGTH},${PUSH_TOKEN_MAX_LENGTH}}$`
-);
+export { PUSH_TOKEN_MIN_LENGTH, PUSH_TOKEN_MAX_LENGTH };
 
 export function normalizePushToken(value: unknown): string | null {
-  if (typeof value !== "string") {
-    return null;
-  }
-
-  const normalized = value.trim();
-  if (!normalized) {
-    return null;
-  }
-
-  return PUSH_TOKEN_FORMAT_REGEX.test(normalized) ? normalized : null;
+  return normalizePushTokenValue(value);
 }
